refactoring of package names to follow Java naming conventions
