--- server/src/main/java/endpoint/Application.java
@@ -1,21 +1,0 @@
-package endpoint;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.web.servlet.ServletComponentScan;
-
-@SpringBootApplication
-// Finds the FhirServlet and runs it
-@ServletComponentScan
-/**
- * Runs the app, including the FhirServlet.  Keep this in the root directory
- * or it won't find the dependencies.  Only initialization functions should be put here.
- */
-public class Application {
-
-
-  public static void main(String[] args) {
-    SpringApplication.run(Application.class, args);
-  }
-
-}
