--- server/src/main/java/endpoint/components/FhirComponents.java
@@ -1,27 +1,0 @@
-package endpoint.components;
-
-import ca.uhn.fhir.context.FhirContext;
-import ca.uhn.fhir.parser.IParser;
-import org.springframework.stereotype.Component;
-
-/**
- * Build some expensive objects here so we can reuse them.
- */
-@Component
-public class FhirComponents {
-  private FhirContext fhirContext;
-  private IParser jsonParser;
-
-  public FhirComponents() {
-    this.fhirContext = FhirContext.forR4();
-    this.jsonParser = fhirContext.newJsonParser();
-  }
-
-  public FhirContext getFhirContext() {
-    return fhirContext;
-  }
-
-  public IParser getJsonParser() {
-    return jsonParser;
-  }
-}
